refactor: move QrCodeButton so a separate fodler
