+++ apps/web/src/components/common/QrCodeButton/QrModal.tsx
