+++ apps/web/src/components/sidebar/SidebarHeader/index.tsx
@@ -16,7 +16,7 @@ import LinkIconBold from '@/public/images/sidebar/link-bold.svg'
 import { selectSettings } from '@/store/settingsSlice'
 import { useCurrentChain } from '@/hooks/useChains'
 import { getBlockExplorerLink } from '@safe-global/utils/utils/chains'
-import QrCodeButton from '../QrCodeButton'
+import QrCodeButton from '@/components/common/QrCodeButton'
 import Track from '@/components/common/Track'
 import { OVERVIEW_EVENTS } from '@/services/analytics/events/overview'
 import { MixpanelEventParams } from '@/services/analytics/mixpanel-events'
