+++ apps/web/src/features/actions-tray/components/ActionsTray/ActionsTray.tsx
@@ -5,7 +5,7 @@ import { ArrowUpRight, ArrowDownLeft, Repeat, SquareDashedBottomCode } from 'luc
 import { Tooltip } from '@mui/material'
 import { Button } from '@/components/ui/button'
 import Track from '@/components/common/Track'
-import QrCodeButton from '@/components/sidebar/QrCodeButton'
+import QrCodeButton from '@/components/common/QrCodeButton'
 import CheckWallet from '@/components/common/CheckWallet'
 import { GeoblockingContext } from '@/components/common/GeoblockingProvider'
 import { AppRoutes } from '@/config/routes'
