+++ apps/web/src/features/spaces/components/Dashboard/AggregatedBalances.tsx
@@ -14,7 +14,7 @@ import { TokenTransferFlow } from '@/components/tx-flow/flows'
 import { MoreVertical } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { DashboardHeader } from '@/features/spaces/components/Dashboard/DashboardHeader'
-import QrModal from '@/components/sidebar/QrCodeButton/QrModal'
+import QrModal from '@/components/common/QrCodeButton/QrModal'
 
 const AggregatedBalance = ({
   safeItems,
