+++ apps/web/src/features/spaces/components/SelectSafeModal/index.tsx
@@ -22,7 +22,7 @@ import { FEATURES, hasFeature } from '@safe-global/utils/utils/chains'
 
 const SWAP_DISABLED_TOOLTIP = 'Swap is not supported on this chain. Try another chain.'
 
-const QrModal = dynamic(() => import('@/components/sidebar/QrCodeButton/QrModal'))
+const QrModal = dynamic(() => import('@/components/common/QrCodeButton/QrModal'))
 
 const SelectSafeModal = () => {
   const [query, setQuery] = useState('')
